% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_machine_learning_fct.R
\name{get_performance}
\alias{get_performance}
\title{Extracts performance from list of caret models}
\usage{
get_performance(fits)
}
\arguments{
\item{fits}{list of caret models}
}
\value{
data frame of accuracy and mcc (classification metrics)
}
\description{
Extracts performance from list of caret models
}
