% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_machine_learning_fct.R
\name{get_roc}
\alias{get_roc}
\title{ROC calculations}
\usage{
get_roc(pred_object, classes)
}
\arguments{
\item{pred_object}{caret model output pred element}
}
\value{
tibble with roc data
}
\description{
ROC calculations
}
