% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{impute}
\alias{impute}
\title{Impute}
\usage{
impute(ms, method)
}
\arguments{
\item{ms}{List object containing value data and rowinfo data (meta data)}

\item{method}{Which imputation method to be used. Includes knn or 0-imputation}
}
\value{
ms object
}
\description{
Imputes missing values. Only used for the app
}
