% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{impute_knn}
\alias{impute_knn}
\title{Missing value imputation: knn replacement}
\usage{
impute_knn(ms)
}
\arguments{
\item{ms}{List object containing value data and rowinfo data (meta data)}
}
\value{
ms
}
\description{
Missing value imputation: knn replacement
}
\examples{
# First convert the pneumonia object to a tibble.
pneumonia<- tibble::tibble(pneumonia)

# Generate list object
ms <- list()

# Assign feature values to ms$values
start_column <- 8 # The first column with feature values
end_column <- ncol(pneumonia) # The last column of the dataset
ms$values <- pneumonia[, start_column:end_column] \%>\% dplyr::slice(1:10)

# Assign metadata to ms$rowinfo
ms$rowinfo <- pneumonia \%>\% dplyr::select(id, group, age, gender, weight, height, BMI) \%>\% dplyr::slice(1:10) \%>\% dplyr::mutate(rowid = dplyr::row_number())
ms <- impute_knn(ms)
head(ms$values)
head(ms$rowinfo)
}
