% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataloader_fct.R
\name{load_ms}
\alias{load_ms}
\title{Convert data frame to ms list object}
\usage{
load_ms(
  df,
  outcome,
  feature_start,
  feature_end = NULL,
  type = NULL,
  tech_rep = NULL,
  covariates = NULL
)
}
\arguments{
\item{df}{data from experiment. Should contain metadata such as outcome, covariates, sample type (e.g., QC, sample, blind) etc.}

\item{outcome}{The outcome}

\item{feature_start}{The first column containing feature/compound values. This function assumes that no metadata columns are between this colum and the final column of feature values}

\item{feature_end}{The final column containing features. Default is last column in df}

\item{type}{If any sample type column exists input it here.}

\item{tech_rep}{If the any column for technical replicates exits input it here.}

\item{covariates}{Add any potential covariates.}
}
\value{
an ms list
}
\description{
Convert data frame to ms list object
}
\examples{
df <- pneumonia
ms <- load_ms(pneumonia, outcome = "group", feature_start = 8)
}
