% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{make_umap}
\alias{make_umap}
\title{UMAP of data (beta)}
\usage{
make_umap(tmp1, tmp2, color_label)
}
\arguments{
\item{tmp1}{metadata}

\item{tmp2}{valuedata}

\item{color_label}{coloring variable}
}
\value{
ggplot2 plot
}
\description{
UMAP of data (beta)
}
\examples{
# First convert the pneumonia object to a tibble.
pneumonia<- tibble::tibble(pneumonia)

# Generate list object
ms <- list()

# Assign feature values to ms$values
start_column <- 8 # The first column with feature values
end_column <- ncol(pneumonia) # The last column of the dataset
ms$values <- pneumonia[, start_column:end_column] \%>\% dplyr::slice(1:100)

# Assign metadata to ms$rowinfo
ms$rowinfo <- pneumonia \%>\% dplyr::select(id, group, age, gender, weight, height, BMI) \%>\% dplyr::slice(1:100) \%>\% dplyr::mutate(rowid = dplyr::row_number())
ms <- impute_zero(ms)
tmp1 <- ms$rowinfo
tmp2 <- ms$values
make_umap(tmp1, tmp2, color_label = "group")
}
