% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_machine_learning_fct.R
\name{mcc}
\alias{mcc}
\title{Matthews Correlation Coefficient}
\usage{
mcc(pred, obs)
}
\arguments{
\item{pred}{predictions from any model}

\item{obs}{observed outcome values}
}
\value{
mcc score
}
\description{
The mcc value is robust to skewed outcome distributions and always represent the amount of information extracted by the model.
}
