% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(ms, method)
}
\arguments{
\item{ms}{List object containing value data and rowinfo data (meta data)}

\item{method}{Which normalization method to be used.}
}
\value{
ms
}
\description{
Normalizes the data. Only used in the app.
}
