% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{normalize_combat}
\alias{normalize_combat}
\title{Normalize using combat}
\usage{
normalize_combat(ms)
}
\arguments{
\item{ms}{List object containing value data and rowinfo data (meta data)}
}
\value{
ms
}
\description{
Normalize using combat
}
