% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{plot_importance}
\alias{plot_importance}
\title{Importance from caret model}
\usage{
plot_importance(model)
}
\arguments{
\item{model}{the caret model trained on the ms list}
}
\value{
ggplot2 plot
}
\description{
Importance from caret model
}
