% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{plot_pca}
\alias{plot_pca}
\title{PCA plot (beta)}
\usage{
plot_pca(ms, color_label = "rowid")
}
\arguments{
\item{ms}{List object containing value and metadata}

\item{color_label}{coloring code for plot}
}
\value{
ggplot2 plot
}
\description{
PCA plot (beta)
}
\examples{
# First convert the pneumonia object to a tibble.
pneumonia<- tibble::tibble(pneumonia)

# Generate list object
ms <- list()

# Assign feature values to ms$values
start_column <- 8 # The first column with feature values
end_column <- ncol(pneumonia) # The last column of the dataset
ms$values <- pneumonia[, start_column:end_column] \%>\% dplyr::slice(1:10)

# Assign metadata to ms$rowinfo
ms$rowinfo <- pneumonia \%>\% dplyr::select(id, group, age, gender, weight, height, BMI) \%>\% dplyr::slice(1:10) \%>\% dplyr::mutate(rowid = dplyr::row_number())
ms <- impute_zero(ms)
plot_pca(ms, color_labels = "rowid")

}
