% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_machine_learning_fct.R
\name{plot_performance}
\alias{plot_performance}
\title{Plot performance of screened ML models}
\usage{
plot_performance(fits)
}
\arguments{
\item{fits}{list of caret models}
}
\value{
ROC plot and repeat-wise performance plot
}
\description{
Plot performance of screened ML models
}
