% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{plot_roc}
\alias{plot_roc}
\title{ROC of caret model}
\usage{
plot_roc(model)
}
\arguments{
\item{model}{the caret model trained on the ms list}
}
\value{
ggplot2 plot
}
\description{
ROC of caret model
}
