% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{plot_standards}
\alias{plot_standards}
\title{Plot intensity vs injection order}
\usage{
plot_standards(ms, standards, batch = NULL)
}
\arguments{
\item{ms}{List object of value and meta data}

\item{standards}{vector containing column names of internal standards
Feature names of standards}

\item{batch}{column name of batch info}
}
\value{
ggplot2 plot
}
\description{
Plot intensity vs injection order
}
\examples{
# First convert the pneumonia object to a tibble.
pneumonia<- tibble::tibble(pneumonia)

# Generate list object
ms <- list()

# Assign feature values to ms$values
start_column <- 8 # The first column with feature values
end_column <- ncol(pneumonia) # The last column of the dataset
ms$values <- pneumonia[, start_column:end_column] \%>\% dplyr::slice(1:10)

# Assign metadata to ms$rowinfo
ms$rowinfo <- pneumonia \%>\% dplyr::select(id, group, age, gender, weight, height, BMI) \%>\% dplyr::slice(1:10) \%>\% dplyr::mutate(rowid = dplyr::row_number())
ms <- impute_zero(ms)
plot_standards(ms, standards = c("M363T419","M512T603","M364T419","M365T392", "M143T177"))
}
