% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pneumonia-data.R
\docType{data}
\name{pneumonia}
\alias{pneumonia}
\title{Untargeted Metabolomics Data}
\format{
An object of class \code{"tbl"}
}
\usage{
pneumonia
}
\description{
Autopsies of pneumonia cases vs controls.
}
\examples{
pneumonia
}
\references{
Elmsjö et al. (2021), DOI: 10.1021/acs.chemrestox.0c00448
}
\keyword{datasets}
