% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{prob_vs_continuous}
\alias{prob_vs_continuous}
\title{Scatter plot of probabilities}
\usage{
prob_vs_continuous(model, target_info, var_name)
}
\arguments{
\item{model}{the caret model trained on the ms list}

\item{target_info}{outcome variable}

\item{var_name}{outcome variable}
}
\description{
Used when classifying over a continuous variable
}
