% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_machine_learning_fct.R
\name{probs_hist}
\alias{probs_hist}
\title{probability histogram of caret model}
\usage{
probs_hist(fits)
}
\arguments{
\item{model}{the caret model trained on the ms list}
}
\value{
ggplot2 plot
}
\description{
probability histogram of caret model
}
