% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{rm_feature_batch_inflated_zeros}
\alias{rm_feature_batch_inflated_zeros}
\title{Remove features inflated with zeros (batch-wise)}
\usage{
rm_feature_batch_inflated_zeros(ms, batch = NULL)
}
\arguments{
\item{ms}{List object containing value data and rowinfo data (meta data)}

\item{batch}{column name of batch. If none provided the dataset is considered as one large batch.}
}
\value{
ms
}
\description{
This might help avoiding batch effect and identification of batches through signals with binary (on/off) behaviour
}
\examples{
# First convert the pneumonia object to a tibble.
pneumonia<- tibble::tibble(pneumonia)
# Generate list object
ms <- list()
# Assign feature values to ms$values
start_column <- 8 # The first column with feature values
end_column <- ncol(pneumonia) # The last column of the dataset
ms$values <- pneumonia[, start_column:end_column] \%>\% dplyr::slice(1:10)
# Assign metadata to ms$rowinfo
ms$rowinfo <- pneumonia \%>\% dplyr::select(id, group, age, gender, weight, height, BMI) \%>\% dplyr::slice(1:10) \%>\% dplyr::mutate(rowid = dplyr::row_number())
ms <- impute_zero(ms)
ms <- rm_feature_batch_inflated_zeros(ms)
head(ms$values)
head(ms$rowinfo)
}
