% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{rm_IS_outliers}
\alias{rm_IS_outliers}
\title{Remove samples if internal standards are outliers}
\usage{
rm_IS_outliers(ms, standards, tolerance = 0, quantiles = c(0.025, 0.975))
}
\arguments{
\item{ms}{{List object containing value data and rowinfo data (meta data)}}

\item{standards}{{Vector containing column names of the internal standards e.g., M215T123}}

\item{tolerance}{{Number of how many internal standards that must lie outside 95 percent of the distribution before a sample is excluded.
Max number of standards
Min 1
The smaller the more restrictive. The tolerance reflects the number of outlier ISs allowed per sample.
Outlier is given by lying in the 5 percent most extreme values. This means that if a threshold of 1 is used 5 percent of samples are removed.
This should be used carefully.}}

\item{quantiles}{{How extreme should an IS value be before being assigned as an outlier.}}
}
\value{
ms
}
\description{
Remove samples if internal standards are outliers
}
