% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{rm_sample_min_batch_occurence}
\alias{rm_sample_min_batch_occurence}
\title{Removes observations if their batches are below a given threshold}
\usage{
rm_sample_min_batch_occurence(ms, min_occurence = 5, batch = NULL)
}
\arguments{
\item{ms}{List object containing value data and rowinfo data (meta data)}

\item{min_occurence}{the minimum number of samples per batch (to avoid batch effect being predictive for the outcome)}

\item{batch}{Column name of the batch info column}
}
\value{
ms
}
\description{
Is usable if preceeded by batch normalization
}
\examples{
# First convert the pneumonia object to a tibble.
pneumonia<- tibble::tibble(pneumonia)

# Generate list object
ms <- list()

# Assign feature values to ms$values
start_column <- 8 # The first column with feature values
end_column <- ncol(pneumonia) # The last column of the dataset
ms$values <- pneumonia[, start_column:end_column] \%>\% dplyr::slice(1:10)

# Assign metadata to ms$rowinfo
ms$rowinfo <- pneumonia \%>\% dplyr::select(id, group, age, gender, weight, height, BMI) \%>\% dplyr::slice(1:10) \%>\% dplyr::mutate(rowid = dplyr::row_number())
ms <- impute_zero(ms)
ms <- rm_sample_min_batch_occurence(ms)
head(ms$values)
head(ms$rowinfo)
}
