% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{rm_sample_pca_outliers}
\alias{rm_sample_pca_outliers}
\title{Removes outliers by PCA}
\usage{
rm_sample_pca_outliers(ms, plot = FALSE)
}
\arguments{
\item{ms}{List object containing value data and rowinfo data (meta data)}

\item{plot}{set to TRUE to visualize the removed samples}
}
\value{
ms
}
\description{
Samples are removed if: abs(x - median(x)) > (1.5 * quantile(x,0.95)-quantile(x,0.05)), where x is the PCA scores.
Scores from the first 12 components are used.
}
\details{
May be applied before row normalization but after removal of extreme features.
}
\examples{
# First convert the pneumonia object to a tibble.
pneumonia<- tibble::tibble(pneumonia)

# Generate list object
ms <- list()

# Assign feature values to ms$values
start_column <- 8 # The first column with feature values
end_column <- ncol(pneumonia) # The last column of the dataset
ms$values <- pneumonia[1:10, start_column:end_column]

# Assign metadata to ms$rowinfo
ms$rowinfo <- pneumonia[1:10,] \%>\% dplyr::select(id, group, age, gender)
ms <- impute_zero(ms)
ms <- rm_sample_pca_outliers(ms)
}
