% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{transform}
\alias{transform}
\title{Transform}
\usage{
transform(ms, method)
}
\arguments{
\item{ms}{List object containing value data and rowinfo data (meta data)}

\item{method}{Which transformation method to be used (includes fourth root and log10)}
}
\value{
ms object
}
\description{
Wrapper for all the transformation methods. Only used in the app.
}
