% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-unbiased_stICA.R
\name{unbiased_stICA}
\alias{unbiased_stICA}
\title{Utility function for waveICA}
\usage{
unbiased_stICA(X, k = 10, alpha)
}
\arguments{
\item{X}{values}

\item{k}{number of components}

\item{alpha}{tradeoff between spatial ICA and temporal ICA}
}
\value{
components
}
\description{
Utility function for waveICA
}
