% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-WaveICA_2.0.R
\name{WaveICA_2.0}
\alias{WaveICA_2.0}
\title{WaveICA2.0}
\usage{
WaveICA_2.0(data, wf = "haar", Injection_Order, alpha, Cutoff, K)
}
\arguments{
\item{data}{Sample-by-Peak matrix untargeted metabolomics data.}

\item{wf}{wavelet functions. The default is "haar".}

\item{Injection_Order}{Injection Order of samples.}

\item{alpha}{The trade-off value between the independence of samples and those
of variables in ICA, and should be between 0 and 1.The default is 0.}

\item{Cutoff}{The threshold of the variation explained by the injection order for
independent components. It should be between 0 and 1.}

\item{K}{The maximal component that ICA decomposes. The default is 10.}
}
\value{
A list that contains the clean data.
}
\description{
Removing batch effects for metabolomics data without
using batch information
}
\author{
Kui Deng
\email{dengkui_stat@163.com}
}
