% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_machine_learning_fct.R
\name{fit_models}
\alias{fit_models}
\title{Fit models
Use several different model types to predict on the x and y}
\usage{
fit_models(x, y, methods = c("glmnet", "pls"))
}
\arguments{
\item{methods}{}
}
\description{
Fit models
Use several different model types to predict on the x and y
}
