% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_machine_learning_fct.R
\name{fit_models}
\alias{fit_models}
\title{Fit models
Use several different model types to predict on the x and y}
\usage{
fit_models(
  ms,
  outcome,
  analysis_type = "Classification",
  methods = c("glmnet", "pls")
)
}
\arguments{
\item{methods}{machine learning models}

\item{x}{value data (matrix, df, or tibble)}

\item{y}{outcome variable (vector)}
}
\value{
list of caret models
}
\description{
Fit models
Use several different model types to predict on the x and y
}
