% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{normalize_pqn}
\alias{normalize_pqn}
\title{Probabilistic quotient normalization}
\usage{
normalize_pqn(ms, n = "median", QC = NULL)
}
\arguments{
\item{ms}{List object containing value data and rowinfo data (meta data)}

\item{n}{Summary function}

\item{QC}{Available QC column?}
}
\value{
ms
}
\description{
Probabilistic quotient normalization
}
\examples{
# First convert the pneumonia object to a tibble.
pneumonia<- tibble::tibble(pneumonia)

# Generate list object
ms <- list()

# Assign feature values to ms$values
start_column <- 8 # The first column with feature values
end_column <- ncol(pneumonia) # The last column of the dataset
ms$values <- pneumonia[1:10, start_column:end_column]

# Assign metadata to ms$rowinfo
ms$rowinfo <- pneumonia[1:10,] \%>\% dplyr::select(rowid = id, group, age, gender)
ms <- impute_zero(ms)
ms <- normalize_pqn(ms)
head(ms$values)
head(ms$rowinfo)
}
