% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{plot_pca}
\alias{plot_pca}
\title{PCA plot (beta)}
\usage{
plot_pca(ms, color_label = "rowid")
}
\arguments{
\item{ms}{List object containing value and metadata}

\item{color_label}{coloring code for plot}
}
\value{
ggplot2 plot
}
\description{
PCA plot (beta)
}
\examples{
# First convert the pneumonia object to a tibble.
pneumonia<- tibble::tibble(pneumonia)

# Generate list object
ms <- list()

# Assign feature values to ms$values
start_column <- 8 # The first column with feature values
end_column <- ncol(pneumonia) # The last column of the dataset
ms$values <- pneumonia[1:100, start_column:end_column]

# Assign metadata to ms$rowinfo
ms$rowinfo <- pneumonia[1:100,] \%>\% dplyr::select(rowid = id, group, age, gender)
ms <- impute_zero(ms)
plot_pca(ms, color_label = "rowid")

}
