% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_machine_learning_fct.R,
%   R/mod_preprocess_fct.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot performance of screened ML models}
\usage{
plot_roc(model)

plot_roc(model)
}
\arguments{
\item{model}{the caret model trained on the ms list}

\item{fits}{list of caret models}
}
\value{
ROC plot and repeat-wise performance plot

ggplot2 plot
}
\description{
Plot performance of screened ML models

ROC of caret model
}
