% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{rm_IS_outliers}
\alias{rm_IS_outliers}
\title{Remove samples if internal standards are outliers}
\usage{
rm_IS_outliers(
  ms,
  standards,
  tolerance = 0,
  quantiles = c(0.025, 0.975),
  batch = NULL
)
}
\arguments{
\item{ms}{{List object containing value data and rowinfo data (meta data)}}

\item{standards}{{Vector containing column names of the internal standards e.g., M215T123}}

\item{tolerance}{{Number of how many internal standards that must lie outside 95 percent of the distribution before a sample is excluded.
Max number of standards
Min 1
The smaller the more restrictive. The tolerance reflects the number of outlier ISs allowed per sample.
Outlier is given by lying in the 5 percent most extreme values. This means that if a threshold of 1 is used 5 percent of samples are removed.
This should be used carefully.}}

\item{quantiles}{{How extreme should an IS value be before being assigned as an outlier.}}

\item{batch}{column name of batch info}
}
\value{
ms
}
\description{
Remove samples if internal standards are outliers
}
\examples{
# First convert the pneumonia object to a tibble.
pneumonia <- tibble::tibble(pneumonia)
# Generate list object
ms <- list()
# Assign feature values to ms$values
start_column <- 8 # The first column with feature values
end_column <- ncol(pneumonia) # The last column of the dataset
ms$values <- pneumonia[1:10, start_column:end_column]
# Assign metadata to ms$rowinfo
ms$rowinfo <- pneumonia \%>\%
    dplyr::select(id, group, age, gender) \%>\%
    dplyr::slice(1:10) \%>\%
    dplyr::mutate(rowid = dplyr::row_number())
ms <- impute_zero(ms)
ms <- rm_IS_outliers(ms, standards=c("M363T419","M512T603"), tolerance=4, quantiles=c(0.01,0.99))

}
