% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{select_most_variable_features}
\alias{select_most_variable_features}
\title{Select most variable features (beta version)
Only works with features starting with "M" as in M314T182 (untargeted data)
Effective for model screening to avoid heavy work loads. Often top 500 contains most of the signal.}
\usage{
select_most_variable_features(ms, n = 500)
}
\arguments{
\item{ms}{List object containing value data and rowinfo data (meta data)}

\item{n}{the number of most variable features selected}
}
\value{
ms
}
\description{
Select most variable features (beta version)
Only works with features starting with "M" as in M314T182 (untargeted data)
Effective for model screening to avoid heavy work loads. Often top 500 contains most of the signal.
}
\examples{
# First convert the pneumonia object to a tibble.
pneumonia<- tibble::tibble(pneumonia)

# Generate list object
ms <- list()

# Assign feature values to ms$values
start_column <- 8 # The first column with feature values
end_column <- ncol(pneumonia) # The last column of the dataset
ms$values <- pneumonia[1:10, start_column:end_column]

# Assign metadata to ms$rowinfo
ms$rowinfo <- pneumonia[1:10,] \%>\% dplyr::select(id, group, age, gender)
ms <- impute_zero(ms)
ms <-select_most_variable_features(ms, n=10)
head(ms$values)
head(ms$rowinfo)

}
