% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess_fct.R
\name{standardize_z_batch}
\alias{standardize_z_batch}
\title{Z-standardization of batches}
\usage{
standardize_z_batch(ms, batch_column = NULL)
}
\arguments{
\item{ms}{List object containing value and meta data}

\item{batch_column}{Column name of batch info}
}
\value{
ms
}
\description{
Z-standardization of batches
}
\examples{
# First convert the pneumonia object to a tibble.
pneumonia<- tibble::tibble(pneumonia)

# Generate list object
ms <- list()

# Assign feature values to ms$values
start_column <- 8 # The first column with feature values
end_column <- ncol(pneumonia) # The last column of the dataset
ms$values <- pneumonia[, start_column:end_column] \%>\% dplyr::slice(1:10)

# Assign metadata to ms$rowinfo
ms$rowinfo <- pneumonia \%>\% dplyr::select(id, group, age, gender, weight, height, BMI) \%>\% dplyr::slice(1:10) \%>\% dplyr::mutate(rowid = dplyr::row_number())
ms <- impute_zero(ms)
ms <- standardize_z_batch(ms)
head(ms$values)
head(ms$rowinfo)

}
